final refactoring
