+++ src/wags_tails/chembl.py
@@ -7,8 +7,8 @@ from typing import Optional
 
 import requests
 
-from wags_tails.base_source import DataSource, RemoteDataError
-from wags_tails.download_utils import download_http
+from .base_source import DataSource, RemoteDataError
+from .core_utils.downloads import download_http
 
 
 class ChemblData(DataSource):
