+++ src/wags_tails/core_utils/__init__.py
@@ -1,0 +1,4 @@
+"""Provide generic helper functions for frequent data-related tasks.
+
+Intended to be largely for internal use.
+"""
