+++ src/wags_tails/drugsatfda.py
@@ -5,9 +5,9 @@ from typing import Optional
 
 import requests
 
-from wags_tails.base_source import DataSource, RemoteDataError
-from wags_tails.download_utils import download_http, handle_zip
-from wags_tails.version_utils import DATE_VERSION_PATTERN
+from .base_source import DataSource, RemoteDataError
+from .core_utils.downloads import download_http, handle_zip
+from .core_utils.versioning import DATE_VERSION_PATTERN
 
 
 class DrugsAtFdaData(DataSource):
