+++ src/wags_tails/hemonc.py
@@ -7,10 +7,10 @@ from typing import NamedTuple, Optional, Tuple
 
 import requests
 
-from wags_tails.base_source import DataSource, RemoteDataError
-from wags_tails.download_utils import download_http
-from wags_tails.storage_utils import get_latest_local_file
-from wags_tails.version_utils import parse_file_version
+from .base_source import DataSource, RemoteDataError
+from .core_utils.downloads import download_http
+from .core_utils.storage import get_latest_local_file
+from .core_utils.versioning import parse_file_version
 
 _logger = logging.getLogger(__name__)
