+++ src/wags_tails/ncit.py
@@ -5,8 +5,8 @@ from typing import Optional
 
 import requests
 
-from wags_tails.base_source import DataSource, RemoteDataError
-from wags_tails.download_utils import download_http, handle_zip
+from .base_source import DataSource, RemoteDataError
+from .core_utils.downloads import download_http, handle_zip
 
 
 class NcitData(DataSource):
